Move statevector latex tests and fix code to pass
